fix Counts reported by the statistics bar do not fully align with counts reported via dashboard tabs
