+++ src/main/frontend/src/utils/dashboardUtilFunctions.js
@@ -186,25 +186,6 @@ export const unselectFilters = (filtersObj) => filtersObj.map((filterElement) =>
   isChecked: false,
 }));
 
-export function getStatDataFromDashboardData(data, statName) {
-  switch (statName) {
-    case 'case':
-      return [...new Set(data.filter((d) => d.case_id).map((d) => d.case_id))].length;
-    case 'study':
-      return [...new Set(data.map((d) => d.study_code))].length;
-    case 'aliquot':
-      return 0;
-    case 'sample':
-      return [...new Set(data.reduce((output, d) => output.concat(d.samples
-        ? d.samples : []), []))].length;
-    case 'file':
-      return [...new Set(data.reduce((output, d) => output.concat(d.files
-        ? d.files : []), []).map((f) => f.uuid))].length;
-    default:
-      return 0;
-  }
-}
-
 // getStudiesProgramWidgetFromDT
 
 export function getSunburstDataFromDashboardData(data) {
@@ -654,4 +635,101 @@ export function formatFileSize(bytes, decimals = 2) {
   const i = Math.floor(Math.log(bytes) / Math.log(1024));
 
   return `${parseFloat((bytes / (1024 ** i)).toFixed(dm))} ${sizes[i]}`;
+}
+
+export function FileData(fileData) {
+  // combine case properties with files.
+  const transform = (accumulator, currentValue) => {
+    const caseAttrs = {};
+    Object.keys(currentValue).forEach((key) => {
+      if (key && !Array.isArray(currentValue[key])) {
+        caseAttrs[key] = currentValue[key];
+      }
+    });
+    if (currentValue.files) {
+      return accumulator.concat(currentValue.files.map((f) => ({ ...f, ...caseAttrs })));
+    }
+    return accumulator;
+  };
+
+  const tableData = fileData.data.reduce(transform, []);
+
+  // reduce duplicated records based on file's uuid
+  const result = [];
+  const map = new Map();
+  tableData.forEach((item) => {
+    if (!map.has(item.uuid)) {
+      map.set(item.uuid, true); // set any value to Map
+      result.push(item);
+    }
+  });
+
+  // get files filters
+  const filesFilters = JSON.parse(JSON.stringify(fileData)).filters
+    .filter((f) => f.section === 'file')
+    .map((f) => {
+      const tmpF = f;
+      tmpF.datafield = tmpF.datafield.includes('@') ? tmpF.datafield.split('@').pop() : tmpF.datafield;
+      return tmpF;
+    });
+
+  // filter out the records which does not match the filters
+  const tableDataAfterFilter = result.filter((row) => filterData(row, filesFilters));
+
+  return tableDataAfterFilter;
+}
+
+export function SampleData(sampleData) {
+  // combine case properties with samples.
+  const transform = (accumulator, currentValue) => {
+    const caseAttrs = {};
+    Object.keys(currentValue).forEach((key) => {
+      if (key && !Array.isArray(currentValue[key])) {
+        caseAttrs[key] = currentValue[key];
+      }
+    });
+    if (currentValue.sample_list) {
+      return accumulator.concat(currentValue.sample_list.map((f) => ({ ...f, ...caseAttrs })));
+    }
+    return accumulator;
+  };
+  const tableData = sampleData.data.reduce(transform, []);
+
+  // get sample filters
+  const sampleFilters = JSON.parse(JSON.stringify(sampleData)).filters
+    .filter((f) => f.section === 'sample')
+    .map((f) => {
+      const tmpF = f;
+      tmpF.datafield = tmpF.datafield.includes('@') ? tmpF.datafield.split('@').pop() : tmpF.datafield;
+      return tmpF;
+    });
+
+  // filter out the records which does not match the filters, plus case_id should not be null.
+  const tableDataAfterFilter = tableData
+    .filter((row) => filterData(row, sampleFilters))
+    .filter((d) => {
+      if (d.case_id) {
+        return true;
+      }
+      return false;
+    });
+
+  return tableDataAfterFilter;
+}
+
+export function getStatDataFromDashboardData(data, statName, filters) {
+  switch (statName) {
+    case 'case':
+      return [...new Set(data.filter((d) => d.case_id).map((d) => d.case_id))].length;
+    case 'study':
+      return [...new Set(data.map((d) => d.study_code))].length;
+    case 'aliquot':
+      return 0;
+    case 'sample':
+      return SampleData({ data, filters }).length;
+    case 'file':
+      return FileData({ data, filters }).length;
+    default:
+      return 0;
+  }
 }
